Build libedit, libncurses if needed.

+++ libedit/build-libedit.sh
@@ -2,6 +2,11 @@
 
 set -eux
 
+if [[ -f /usr/local/include/editline/readline.h && -f /usr/local/lib/libedit.a ]]; then
+  echo Found libedit.
+  exit 0
+fi
+
 ver=20210910-3.1
 [ -d libedit-${ver} ] && rm -rf libedit-${ver}
 tar xvf libedit-${ver}.tar.xz
